obsservice: move the http proxy to its own package
Release note: None
